third_party/blink/web_tests/http/tests/websocket/multiple-connections-throttled-expected.txt

Test that WebSocket connections are throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS openedCount is 255
PASS successfullyParsed is true

TEST COMPLETE
